pyproject.toml
[project]
name = "cli-llm"
version = "2025.7.0"
description = "A DSPy-powered boilerplate that boosts LLM usability in the CLI with customizable, high-performance features."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dspy>=2.6.19",
    "pre-commit>=4.2.0",
    "python-dotenv>=1.1.0",
    "rich>=14.0.0",
    "ruff>=0.11.7",
]
